Refactor client.py
